requirements.txt
streamlit
streamlit-lottie
requests
pandas
qrcode
reportlab
pushbullet.py
scikit-learn
plotly